Added support for revoking KeyCloak tokens
The revocation endpoint was forced to be null, when using the keycloak oauth2 provider. 

As per the keycloak documentation, there is an endpoint for revoking tokens: https://www.keycloak.org/docs/latest/securing_apps/#_token_revocation_endpoint

This would enable logging users out with API instead of redirects with keycloak.