Refactor JWT signing algorithm creation
Motivation:

JWT requires the creation of signing algorithm to use for the creation and validation of tokens.

This could be used to implement Quic token validation process and thus we could reuse the implementation engine to implement a Quic token validator.

This refactor also ease the creation usage of thread local to replace synchronised blocks around java cryptographic objet non thread safe usage.

Changes:

Extract the sign/verify part into a small package independent of JWT that can be reused in this module.